requirements.txt
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
gradio>=4.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
